+++ tests/test_temp.py
@@ -1,6 +1,0 @@
-import unittest
-
-class TestTemp(unittest.TestCase):
-
-    def test_temp_func(self):
-        self.assertTrue(True)