+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
-## unreleased
+## (v0.2.1rc2) (-/-/2023)
 
 ### Added
+- Added functionality to define an area (Circle, Envelope)  
+- Added functionality to communicate with BRO REST API and retrieve CPT availability in area
+- Added parsing functionality of the Characteristics
+- Added parsing functionality of the returned CPT XML files from the BRO
+- Added functionality to get the available CPTs within the requested area in GeoJSON format
+- Added tests
 
 ### Changed